+++ filesel/filesystem-bzip2-test.c
@@ -96,12 +96,12 @@ void register_dirdecompressor (const struct ocpdirdecompressor_t *ofd)
 {
 }
 
-int adbMetaAdd (const char *filename, const size_t filesize, const char *SIG, const unsigned char  *data, const size_t  datasize)
+int adbMetaAdd (const char *filename, const uint64_t filesize, const char *SIG, const unsigned char  *data, const uint32_t  datasize)
 {
 	return 0;
 }
 
-int adbMetaGet (const char *filename, const size_t filesize, const char *SIG,       unsigned char **data,       size_t *datasize)
+int adbMetaGet (const char *filename, const uint64_t filesize, const char *SIG,       unsigned char **data,       uint32_t *datasize)
 {
 	return -1;
 }
